# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PicHub-Aggregator 图片聚合分发服务。管理接口支持会话和 Bearer token 鉴权，图片分发接口公开。",
        "title": "PicHub Aggregator API",
        "termsOfService": "https://github.com/untitled572/pichub-aggregator",
        "contact": {
            "name": "PicHub"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.6.5"
    },
    "basePath": "/",
    "paths": {
...
        "/api/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理接口，不回显 admin_token / admin_password_hash",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "读取设置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Settings"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "admin_password 留空表示不修改；admin_token 留空保持现有令牌。凭据变更会全量踢下线",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "更新设置",
                "parameters": [
                    {
                        "description": "设置项",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Settings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Settings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/e/{name}": {
            "get": {
                "description": "?category 为空时 fallback 到端点自身绑定的标签",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Distribution"
                ],
                "summary": "自定义端点分发",
                "parameters": [
                    {
                        "type": "string",
                        "description": "端点名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标签(可逗号分隔多个)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "json"
                        ],
                        "type": "string",
                        "description": "响应格式",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "mobile",
                            "desktop"
                        ],
                        "type": "string",
                        "description": "根据 UA 推断的屏幕方向",
                        "name": "orientation",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.Result"
                        }
                    },
                    "302": {
                        "description": "重定向到图片 URL (format != json)"
                    },
                    "404": {
                        "description": "端点不存在或已禁用",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "无可用图片",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/random": {
            "get": {
                "description": "从分发池随机返回一张图片。category 为空时使用 settings.bound_tags。format=json 返回 JSON，否则 302 重定向到图片 URL。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Distribution"
                ],
                "summary": "随机分发图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "标签(可逗号分隔多个)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "json"
                        ],
                        "type": "string",
                        "description": "响应格式",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "mobile",
                            "desktop"
                        ],
                        "type": "string",
                        "description": "根据 UA 推断的屏幕方向",
                        "name": "orientation",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.Result"
                        }
                    },
                    "302": {
                        "description": "重定向到图片 URL (format != json)"
                    },
                    "503": {
                        "description": "无可用图片",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/random/detect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "抓取目标 URL 并推断响应类型、headers 与可能的图片 URL 路径",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Distribution"
                ],
                "summary": "探测数据源结构",
                "parameters": [
                    {
                        "description": "目标 URL",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DetectURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.DetectResult"
                        }
                    },
                    "400": {
                        "description": "请求体无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "探测失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.DetectURLRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.Settings": {
            "type": "object",
            "properties": {
                "admin_password": {
                    "type": "string"
                },
                "admin_username": {
                    "type": "string"
                },
                "bound_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "cache_max_images": {
                    "type": "integer"
                },
                "cache_max_mb": {
                    "type": "integer"
                },
                "cache_ttl": {
                    "type": "integer"
                },
                "health_check_interval": {
                    "type": "integer"
                },
                "login_enabled": {
                    "type": "boolean"
                },
                "max_history_records": {
                    "type": "integer"
                },
                "min_resolution": {
                    "type": "string"
                },
                "pool_size": {
                    "type": "integer"
                },
                "precache_count": {
                    "type": "integer"
                },
                "proxy_enabled": {
                    "type": "boolean"
                },
                "proxy_mode": {
                    "type": "boolean"
                },
                "proxy_url": {
                    "type": "string"
                },
                "rate_limit": {
                    "type": "integer"
                },
                "rate_limit_window": {
                    "type": "integer"
                },
                "saved_images_dir": {
                    "type": "string"
                },
                "session_hours": {
                    "type": "integer"
                },
                "timeout": {
                    "type": "integer"
                }
            }
        },
...
        "service.DetectResult": {
            "type": "object",
            "properties": {
                "body_tree": {},
                "error": {
                    "type": "string"
                },
                "final_url": {
                    "type": "string"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "resp_type": {
                    "type": "string"
                },
                "url_hints": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "service.Result": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "file_id": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "image_id": {
                    "type": "integer"
                },
                "local_url": {
                    "type": "string"
                },
                "orientation": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "source_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}